Docker-Compose-Project/backend/requirements.txt
Flask 
mysql-connector-python